:label: arrange types #100
