+++ frontend/src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import { configureStore } from "@reduxjs/toolkit";
 import { Provider } from "react-redux";
-import rootReducer from "./redux";
+import rootReducer from "./status";
 
 const store = configureStore({reducer: rootReducer});
 store.subscribe(()=>{console.log("subscribe = ", store.getState())});
