+++ frontend/src/pages/Main.tsx
@@ -1,15 +1,5 @@
 import "./main.css";
 
-export type userInfo = {
-  user_id: number;
-  intra_id: string;
-  auth?: boolean;
-  email: string;
-  phone?: string;
-  access: string;
-  refresh: string;
-};
-
 export default function Main() {
   const url: string = "/auth/login";
