+++ frontend/src/types/userTypes.tsx
@@ -1,0 +1,9 @@
+export type userInfo = {
+  user_id: number;
+  intra_id: string;
+  auth?: boolean;
+  email: string;
+  phone?: string;
+  access: string;
+  refresh: string;
+};
